pyproject.toml
[project]
name = "vissibl"
version = "0.1.0"
description = "FastAPI backend for Vissibl. Project Dubai aims to provide ISO 9001:2015 compliant documentation for the companies in 10 minutes or less."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "faker>=33.1.0",
    "fastapi[all,standard]>=0.115.5",
    "httpx>=0.28.0",
    "ipython>=8.30.0",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "psycopg2>=2.9.10",
    "pycountry>=24.6.1",
    "pydantic-extra-types>=2.10.0",
    "pydantic-settings>=2.6.1",
    "pymupdf4llm>=0.0.17",
    "pymupdf>=1.24.14",
    "pytest>=8.3.4",
    "python-slugify[unidecode]>=8.0.4",
    "sqlalchemy[asyncio]>=2.0.36",
    "sqlmodel>=0.0.22",
    "typer>=0.14.0",
    "typing-extensions>=4.12.2",
    "google-generativeai>=0.8.3",
    "jinja2>=3.1.4",
    "langchain>=0.3.10",
    "langchain-core>=0.3.22",
    "langchain-community>=0.3.10",
    "langchain-google-genai>=2.0.6",
    "pinecone[grpc]>=5.4.2",
    "google-cloud-aiplatform>=1.74.0",
    "psycopg2-binary>=2.9.10",
    "pydantic==2.9",
    "resend>=2.4.0",
    "faiss-cpu>=1.9.0.post1",
    "langchain-openai>=0.2.12",
    "weasyprint>=63.1",
    "markdown>=3.7",
    "beautifulsoup4>=4.12.3",
    "shortuuid>=1.0.13",
    "uvicorn>=0.32.1",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.4",
]

[tool.semantic_release]
version_variables = ["app/version.py:__version__"]
